+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Summary in euro (via filter)
 -   Filter summary by ticker
 
+### Fixed
+
+-   Dividend tax calculation
+
 ## [v0.4.0] - 2022-03-22
 
 ### Added